feat(v12): multitenancy plane, support sessions, and the explorer on the store
A checkpoint of the v12 work in flight, green on `go build ./...` and the
backend test suite, taken before the domain layer is reworked.

Tenancy plane
- /tenants for the instance operator: cards per customer with live counts,
  limits, and create/edit/terminate. Hidden and route-gated behind the
  platform identity *and* an MSSP licence — without MSSP the single tenant's
  admin is the platform identity, so the platform check alone let them reach
  a page the API answers 403 to.
- Settings -> Support access for the customer: NONE / READ / FULL, granted by
  them and never by us. Adds GET /tenants/:id/support-access, because a tenant
  could set the grant but not read it back.
- Support sessions: entering a tenant stamps X-Tenant-Id on every call except
  /auth/* and /tenants. Blanket stamping logs the operator out — /auth/me
  resolves their own account, which lives in the platform tenant, and the 401
  drives the client's refresh-then-logout path.

Fixes found by exercising the above
- appconfig: Order(gorm.Expr(...)) is silently discarded by GORM (Order takes
  only a string or a clause.OrderByColumn), so the inheritance query ran
  unordered and a tenant read the platform's config back over its own. Moved
  the choice into Go where the compiler can see it.
- tenant: a per-tenant AI cap was compared against an instance limit whose
  default is -1, so every cap was thrown away and "deny" meant "unlimited".
- socai: the default-provisioning loop ran in every replica; now claimed
  through joblease.
- threatintel: /usage reported the whole instance's consumption to any tenant.
  Platform-only now, with a /status endpoint for the "is it configured" probe
  it was being used for.
- loganalyzer: the update path wrote the whole row from a body that carries no
  tenant, blanking it — and since reads are scoped, saving a query made it
  vanish for its owner.

Log explorer on the event store
- Documents now come from POST /log-analyzer/search rather than the OpenSearch
  gateway, with the window in the scope so the store narrows by time first.
- The selector picks a dataType, not an index pattern: the explorer always
  reads logs, and the types come from the data so one that stopped arriving
  stops being offered.
- Saved searches moved from localStorage to the tenant-scoped table that
  already existed unused, with a one-time import of what a browser held.
- utm_log_analyzer_query -> saved_query, la_* columns dropped.

Still on the old contract: the explorer's SQL mode and CSV export, so
OpenSearch cannot be retired yet.

go.mod no longer carries a local-path replace for the SDK; it pins the pushed
commit instead, so this builds anywhere.

+++ backend/modules/compliance/handler/report_schedule.go
@@ -37,7 +37,7 @@ func (h *ScheduleHandler) Create(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
-	resp, err := h.uc.Create(c.Request.Context(), int64(actor), req)
+	resp, err := h.uc.Create(c.Request.Context(), actor, req)
 	ev := audit_connectors.Event{Action: "compliance.schedule.create", ResourceType: "compliance_report_schedule"}
 	if resp != nil {
 		ev.ResourceID = strconv.FormatInt(resp.ID, 10)
@@ -70,7 +70,7 @@ func (h *ScheduleHandler) Update(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
-	resp, err := h.uc.Update(c.Request.Context(), int64(actor), req)
+	resp, err := h.uc.Update(c.Request.Context(), actor, req)
 	audit.Record(c, audit_connectors.Event{Action: "compliance.schedule.update", ResourceType: "compliance_report_schedule", ResourceID: strconv.FormatInt(req.ID, 10)},
 		audit_domain.COMPLIANCE_SCHEDULE_UPDATE_ATTEMPT, audit_domain.COMPLIANCE_SCHEDULE_UPDATE_SUCCESS, err)
 	if err != nil {
@@ -98,7 +98,7 @@ func (h *ScheduleHandler) ListByUser(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
-	items, total, err := h.uc.ListByUser(c.Request.Context(), int64(actor), f)
+	items, total, err := h.uc.ListByUser(c.Request.Context(), actor, f)
 	if err != nil {
 		writeError(c, err)
 		return
